Fix for when content_length isn't set.
When a 304, for example, is returned there is no content and hence no content_length
so I've added a check for content_length being nil. I've also
added a check in unbind for content_length being nil. If I do not do
this the errback is called even though the response parsed correctly.

I'm not sure if this is the correct way to fix this problem but the
existing tests all pass and the (possibly dodgy) sites I've been testing
now work.