set content-length header even on empty POSTS
when posting, it is not required to set a content-length if the length
is zero, but it's recommended to set it wherever the true lenght is
known.  if we know there is no body, we should still set it.  not
setting it currently breaks rails (or thin infront of rails, i'm not
sure which).