Prevent data corruption when decoding gzip stream, fixes #284
This fixes an issue where a gzip stream chunked into smaller pieces
would result in a buffer overrun when decoding the headers.

Previously the position where the headers end in `compressed` was
determined by the delta of `@pos` before and after reading the
headers. This didn't take into account the fact that `@data` may already
contain data, but not enough to have advanced `@pos` on previous
iterations.