Fix for LetsEncrypt root expiry (#352)
Every LetsEncrypt issued signature chain now starts with an expired certificate,
but the second item in the chain is a trusted root.  So instead of failing the
whole validation for any link in the chain failing, just don't add failed links
to the store, then make sure the final certificate is valid given whatever *was*
added to the store.