Remove requires_connection test helper
The aim of this helper is to cause some tests to be skipped if there is
no Internet connection. For some reason (which I haven’t investigated)
when run on GitHub Actions these tests are always being skipped, causing
me to not notice the test failures fixed in 139a59b.

I am happy to remove this check; presumably if there’s no Internet
connection then the affected tests will just fail, and that’s fine by
me. I understand the intent of the author in adding this helper, and
understand that being able to run some of the test suite offline may be
a desirable thing, but I think not at the cost of it hiding failing
tests by default.

(There’s also a requires_port helper which is only used in the
proxy-related tests; in order to remove that one we’d need to start
running proxies locally and in CI, and I don’t want to spend time on
that now. We don’t make use of this library’s proxy functionality in
ably-ruby.)